refactor: introduce non-monobehaviour core classes
