+++ Assets/Scripts/Pointer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00821e6b6cf94085add3f44c5407043c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
