+++ Assets/Scripts/PointerScript.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2020 The Tilt Brush Authors
+// Copyright 2020 The Tilt Brush Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,294 +12,115 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
 namespace TiltBrush
 {
-
     public class PointerScript : MonoBehaviour
     {
-        // ---- public types
+        private readonly Pointer m_Pointer = new Pointer();
 
-        public bool DrawingEnabled;
-        private bool m_WasDrawingEnabled;
-        private CanvasScript m_Canvas;
+        public bool DrawingEnabled
+        {
+            get => m_Pointer.DrawingEnabled;
+            set => m_Pointer.DrawingEnabled = value;
+        }
+
         public CanvasScript Canvas
         {
-            get => m_Canvas;
-            set => m_Canvas = value;
+            get => m_Pointer.Canvas;
+            set => m_Pointer.Canvas = value;
         }
 
-        // ---- Private inspector data
-
-        // ---- Private member data
-
-        public Color m_CurrentColor;
-        public BrushDescriptor m_CurrentBrush;
-        public float m_CurrentBrushSize; // In pointer aka room space
-        private Vector2 m_BrushSizeRange;
-        public float m_CurrentPressure; // TODO: remove and query line instead?
-        private BaseBrushScript m_CurrentLine;
-        private List<ControlPoint> m_ControlPoints;
-        private bool m_LastControlPointIsKeeper;
-
-        // ---- Public properties, accessors, events
-
-        float _FromRadius(float x)
+        public Color m_CurrentColor
         {
-            return Mathf.Sqrt(x);
-        }
-        float _ToRadius(float x)
-        {
-            return x * x;
+            get => m_Pointer.m_CurrentColor;
+            set => m_Pointer.m_CurrentColor = value;
         }
 
-        /// The brush size, using "normalized" values in the range [0,1].
-        /// On get, values are raw and may be outside [0,1].
-        /// On set, values outside of the range [0,1] are clamped.
+        public BrushDescriptor m_CurrentBrush
+        {
+            get => m_Pointer.m_CurrentBrush;
+            set => m_Pointer.m_CurrentBrush = value;
+        }
+
+        public float m_CurrentBrushSize
+        {
+            get => m_Pointer.m_CurrentBrushSize;
+            set => m_Pointer.m_CurrentBrushSize = value;
+        }
+
+        public float m_CurrentPressure
+        {
+            get => m_Pointer.m_CurrentPressure;
+            set => m_Pointer.m_CurrentPressure = value;
+        }
+
         public float BrushSize01
         {
-            get
-            {
-                float min = _FromRadius(m_BrushSizeRange.x);
-                float max = _FromRadius(m_BrushSizeRange.y);
-                return Mathf.InverseLerp(min, max, _FromRadius(BrushSizeAbsolute));
-            }
-            set
-            {
-                float min = _FromRadius(m_BrushSizeRange.x);
-                float max = _FromRadius(m_BrushSizeRange.y);
-                BrushSizeAbsolute = _ToRadius(Mathf.Lerp(min, max, Mathf.Clamp01(value)));
-            }
+            get => m_Pointer.BrushSize01;
+            set => m_Pointer.BrushSize01 = value;
         }
 
-        /// The brush size, in absolute room space units.
-        /// On get, values are raw and may be outside the brush's desired range.
-        /// On set, values outside the brush's nominal range are clamped.
         public float BrushSizeAbsolute
         {
-            get => m_CurrentBrushSize;
-            set => _SetBrushSizeAbsolute(Mathf.Clamp(value, m_BrushSizeRange.x, m_BrushSizeRange.y));
+            get => m_Pointer.BrushSizeAbsolute;
+            set => m_Pointer.BrushSizeAbsolute = value;
         }
-
-        // ---- Unity events
 
         void Awake()
         {
-            m_ControlPoints = new List<ControlPoint>();
-            m_CurrentBrushSize = 1.0f;
-            m_BrushSizeRange.x = 1.0f;
-            m_BrushSizeRange.y = 2.0f;
-            m_CurrentPressure = 1.0f;
+            m_Pointer.Initialize();
         }
 
         void Update()
         {
-            if (DrawingEnabled && !m_WasDrawingEnabled)
-            {
-                // Drawing just got enabled, so we need to create a new line.
-                // This is a no-op if the current line is already set.
-                CreateNewLine(m_Canvas, TrTransform.FromLocalTransform(transform));
-            }
-            else if (DrawingEnabled && m_WasDrawingEnabled)
-            {
-                UpdateLineFromObject();
-            }
-            else if (!DrawingEnabled && m_WasDrawingEnabled)
-            {
-                DetachLine(false);
-            }
-            m_WasDrawingEnabled = DrawingEnabled;
+            m_Pointer.Tick(transform);
         }
 
-        /// Returns xf_RS, relative to the passed line transform.
-        /// Applies m_LineDepth and ignores xf_RS.scale
-        /// TODO: see above.
-        TrTransform GetTransformForLine(Transform line, TrTransform xf_RS)
-        {
-            var xfRoomFromLine = Coords.AsRoom[line];
-            xf_RS.translation += xf_RS.forward;
-            xf_RS.scale = 1;
-            return TrTransform.InvMul(xfRoomFromLine, xf_RS);
-        }
-
-        /// Non-playback case:
-        /// - Update the stroke based on the object's position.
-        /// - Save off control points
-        /// - Play audio.
         public void UpdateLineFromObject()
         {
-            if (m_CurrentLine == null) return;
-            var xf_LS = GetTransformForLine(m_CurrentLine.transform, Coords.AsRoom[transform]);
-
-            bool bQuadCreated = m_CurrentLine.UpdatePosition_LS(xf_LS, m_CurrentPressure);
-
-            // TODO: let brush take care of storing control points, not us
-            SetControlPoint(xf_LS, isKeeper: bQuadCreated);
-            UpdateLineVisuals();
+            m_Pointer.UpdateLineFromObject(transform);
         }
 
-        /// Playback case:
-        /// - Update stroke based on the passed transform (in local coordinates)
-        /// - Do _not_ apply any normal adjustment; it's baked into the control point
-        /// - Do not update the mesh
-        /// TODO: replace with a bulk-ControlPoint API
         public void UpdateLineFromControlPoint(ControlPoint cp)
         {
-            float scale = m_CurrentLine.StrokeScale;
-            m_CurrentLine.UpdatePosition_LS(
-                TrTransform.TRS(cp.m_Pos, cp.m_Orient, scale), cp.m_Pressure);
+            m_Pointer.UpdateLineFromControlPoint(cp);
         }
 
-        /// Bulk control point addition
         public void UpdateLineFromStroke(Stroke stroke)
         {
-            float scale = m_CurrentLine.StrokeScale;
-            foreach (var cp in stroke.m_ControlPoints.Where((x, i) => !stroke.m_ControlPointsToDrop[i]))
-            {
-                m_CurrentLine.UpdatePosition_LS(TrTransform.TRS(cp.m_Pos, cp.m_Orient, scale), cp.m_Pressure);
-            }
+            m_Pointer.UpdateLineFromStroke(stroke);
         }
 
         public void UpdateLineVisuals()
         {
-            m_CurrentLine.ApplyChangesToVisuals();
+            m_Pointer.UpdateLineVisuals();
         }
 
-        void _SetBrushSizeAbsolute(float value)
-        {
-            m_CurrentBrushSize = value;
-        }
-
-        /// Pass a Canvas parent, and a transform in that canvas's space.
-        /// If overrideDesc passed, use that for the visuals -- m_CurrentBrush does not change.
         public void CreateNewLine(CanvasScript canvas, TrTransform xf_CS, BrushDescriptor overrideDesc = null)
         {
-            // If straightedge is enabled, we may have a minimum size requirement.
-            // Initialize parametric stroke creator for our type of straightedge.
-            // Maybe change the brush to a proxy brush.
-            BrushDescriptor desc = overrideDesc != null ? overrideDesc : m_CurrentBrush;
-
-            m_CurrentLine = BaseBrushScript.Create(
-                canvas.transform, xf_CS,
-                desc, m_CurrentColor, m_CurrentBrushSize);
+            m_Pointer.CreateNewLine(canvas, xf_CS, overrideDesc);
         }
 
-        /// Like BeginLineFromMemory + EndLineFromMemory
-        /// To help catch bugs in higher-level stroke code, it is considered
-        /// an error unless the stroke is in state NotCreated.
+        public void SetControlPoint(TrTransform lastSpawnXf_LS, bool isKeeper)
+        {
+            m_Pointer.SetControlPoint(lastSpawnXf_LS, isKeeper);
+        }
+
+        public void DetachLine(bool bDiscard)
+        {
+            m_Pointer.DetachLine(bDiscard);
+        }
+
         public void RecreateLineFromMemory(Stroke stroke)
         {
-            if (stroke.m_Type != Stroke.Type.NotCreated)
-            {
-                throw new InvalidOperationException();
-            }
-            if (BeginLineFromMemory(stroke, stroke.Canvas) == null)
-            {
-                // Unclear why it would have failed, but okay.
-                // I guess we keep the old version?
-                Debug.LogError("Unexpected error recreating line");
-                return;
-            }
-
-            UpdateLineFromStroke(stroke);
-
-            // It's kind of warty that this needs to happen; brushes should probably track
-            // the mesh-dirty state and flush it in Finalize().
-            // TODO: Check if this is still necessary now that QuadStripBrushStretchUV
-            // flushes pending geometry changes in Finalize*Brush()
-            m_CurrentLine.ApplyChangesToVisuals();
-
-            // Copy in new contents
-            {
-                m_CurrentLine.FinalizeSolitaryBrush();
-
-                stroke.m_Type = Stroke.Type.BrushStroke;
-                stroke.m_IntendedCanvas = null;
-                stroke.m_Object = m_CurrentLine.gameObject;
-                stroke.m_Object.GetComponent<BaseBrushScript>().Stroke = stroke;
-            }
-
-            m_CurrentLine = null;
+            m_Pointer.RecreateLineFromMemory(stroke, transform);
         }
 
         public GameObject BeginLineFromMemory(Stroke stroke, CanvasScript canvas)
         {
-            BrushDescriptor rBrush = BrushCatalog.GetBrush(stroke.m_BrushGuid);
-            if (rBrush == null)
-            {
-                // Ignore stroke
-                return null;
-            }
-
-            var cp0 = stroke.m_ControlPoints[0];
-            var xf_CS = TrTransform.TRS(cp0.m_Pos, cp0.m_Orient, stroke.m_BrushScale);
-            var xf_RS = canvas.Pose * xf_CS;
-
-            // This transform used to be incorrect, but we didn't notice.
-            // That implies this isn't necessary?
-            transform.position = xf_RS.translation;
-            transform.rotation = xf_RS.rotation;
-
-            m_CurrentBrush = rBrush;
-            m_CurrentBrushSize = stroke.m_BrushSize;
-            m_CurrentColor = stroke.m_Color;
-            CreateNewLine(canvas, xf_CS);
-            m_CurrentLine.SetIsLoading();
-            m_CurrentLine.RandomSeed = stroke.m_Seed;
-
-            return m_CurrentLine.gameObject;
-        }
-
-        /// Record the tranform as a control point.
-        /// If the most-recent point is a keeper, append a new control point.
-        /// Otherwise, the most-recent point is a keeper, and will be overwritten.
-        ///
-        /// The parameter "keep" specifies whether the newly-written point is a keeper.
-        ///
-        /// The current pointer is /not/ queried to get the transform of the new
-        /// control point. Instead, caller is responsible for passing in the same
-        /// xf that was passed to line.UpdatePosition_LS()
-        public void SetControlPoint(TrTransform lastSpawnXf_LS, bool isKeeper)
-        {
-            ControlPoint rControlPoint;
-            rControlPoint.m_Pos = lastSpawnXf_LS.translation;
-            rControlPoint.m_Orient = lastSpawnXf_LS.rotation;
-            rControlPoint.m_Pressure = m_CurrentPressure;
-            rControlPoint.m_TimestampMs = (uint)(App.CurrentSketchTime * 1000);
-
-            if (m_ControlPoints.Count == 0 || m_LastControlPointIsKeeper)
-            {
-                m_ControlPoints.Add(rControlPoint);
-            }
-            else
-            {
-                m_ControlPoints[m_ControlPoints.Count - 1] = rControlPoint;
-            }
-
-            m_LastControlPointIsKeeper = isKeeper;
-        }
-
-
-        // During playback, rMemoryObjectForPlayback is non-null, and strokeFlags should not be passed.
-        // otherwise, rMemoryObjectForPlayback is null, and strokeFlags should be valid.
-        // When non-null, rMemoryObjectForPlayback corresponds to the current line.
-        public void DetachLine(bool bDiscard)
-        {
-            if (bDiscard)
-            {
-                m_CurrentLine.DestroyMesh();
-                Destroy(m_CurrentLine.gameObject);
-            }
-            else
-            {
-                //copy master brush over to current line
-                m_CurrentLine.FinalizeSolitaryBrush();
-            }
-            m_CurrentLine = null;
+            return m_Pointer.BeginLineFromMemory(stroke, canvas, transform);
         }
     }
-} // namespace TiltBrush
+}
